The class statements were modified to inherit from object.
changes to __init__.py:

The set method was modified so it only sets an attribute if the value
is not None.  A reset method was added to set these attributes to None.
The scaling in the _bscaleImage method was changed to omit the offset
of 1, which caused a problem with images that had a range of values
much less than one (e.g. flux calibrated data).

changes to displaydev.py:

A readInfo method was added, in order to get the values of tx, ty,
fbwidth and fbheight.

See also Trac ticket #148.



git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk/numdisplay@1064 fe389314-cf27-0410-b35b-8c050e845b92